API integration tests with jest implemented and UI equipments table with data.

+++ keepit.ui/src/app/core/interceptors/error.interceptor.ts
@@ -1,17 +1,22 @@
 import { HttpInterceptorFn } from '@angular/common/http';
-import { inject } from '@angular/core';
+import { inject, PLATFORM_ID } from '@angular/core';
+import { isPlatformBrowser } from '@angular/common';
 import { catchError, throwError } from 'rxjs';
 import { Router } from '@angular/router';
 
 export const errorInterceptor: HttpInterceptorFn = (req, next) => {
+  const platformId = inject(PLATFORM_ID);
   const router = inject(Router);
 
   return next(req).pipe(
     catchError((error) => {
       if (error.status === 401) {
-        localStorage.clear();
-        router.navigate(['/authentication']);
+        if (isPlatformBrowser(platformId)) {
+          localStorage.clear();
+          router.navigate(['/authentication']);
+        }
       }
+
       return throwError(() => error);
     }),
   );
